+++ example/lib/pages/player/simple/human.dart
@@ -10,11 +10,4 @@ class HumanPlayer extends SimplePlayer {
           size: Vector2.all(24),
           speed: 32,
         );
-
-  @override
-  Future<void> onLoad() {
-    /// Adds rectangle collision
-    add(RectangleHitbox(size: size / 2, position: size / 4));
-    return super.onLoad();
-  }
 }