+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Updade Flame to `1.9.1`
 - rename `bonfireCamera`to `camera`
 - enemy movements improvements
+- Mixin `Vision` improvements
 
 # 3.0.0-beta.17